fix(stats): widen ray-count types to 64-bit (Windows u32 overflow)
Windows-only status-bar ray-count rollover above 2^32: LUMICE_StatsResult
fields and config ray_num used `unsigned long`, which is 32-bit on Windows
(LLP64), silently truncating the size_t totals the core accumulates. Audit
found the whole truncation class, not just the status bar: the stats display
chain, the config ray_num parse/write path, and CLI bench locals.

- Add `LUMICE_RayCount` (unsigned long long) typedef + header-level
  static_assert so every TU compiling lumice.h fails on Windows if it regresses
- Widen LUMICE_StatsResult fields + config ray_num to LUMICE_RayCount
  (ABI change; confirmed no external consumers, GUI/CLI only)
- Widen GUI poller/state stat mirrors + CLI bench locals to LUMICE_RayCount;
  internal threshold/diagnostic cluster to plain unsigned long long
- Regression: field-level static_asserts + >2^32 round-trip tests for config
  ray_num (LUMICE_ParseConfigString) and GUI stats snapshot (ApplyTo)

+++ src/gui/app.cpp
@@ -623,7 +623,7 @@ void CalibrateQualityThreshold() {
   // Uses kCalibrationWindowMs (fixed) instead of kCommitIntervalMs so that changing
   // the commit interval doesn't accidentally tighten/loosen the quality gate.
   double rays_per_window = static_cast<double>(stats[0].sim_ray_num) * kCalibrationWindowMs / elapsed_ms;
-  auto threshold = std::max(kMinRaysFloor, static_cast<unsigned long>(rays_per_window * kCalibrationFraction));
+  auto threshold = std::max(kMinRaysFloor, static_cast<unsigned long long>(rays_per_window * kCalibrationFraction));
 
   g_server_poller.SetCalibratedThreshold(threshold);
   GUI_LOG_INFO("[Calibration] done in {:.0f}ms: {} rays, {:.0f} rays/window({}ms), threshold={}", elapsed_ms,
